+++ news_and_events/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class NewsAndEventsConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "news_and_events"
